+++ include/inc_act/act_structure.php
@@ -98,6 +98,7 @@ if($_SESSION["wcs_user_admin"] == 1) { //Wenn Benutzer Admin-Rechte hat
 		$sql .= "\$indexpage['acat_class']		= '".	str_replace("'", "\\'", $acat_class)."';\n";
 		$sql .= "\$indexpage['acat_keywords']	= '".	str_replace("'", "\\'", $acat_keywords)."';\n";
 		$sql .= "\$indexpage['acat_cpdefault']	= ".	intval($_POST["acat_cpdefault"]).";\n";
+		$sql .= "\$indexpage['acat_disable301']	= ".	(empty($_POST["acat_disable301"]) ? 0 : 1).";\n";
 		$sql .= "?>";
 		write_textfile(PHPWCMS_ROOT.'/config/phpwcms/conf.indexpage.inc.php', $sql);
 	}
@@ -122,7 +123,7 @@ if($_SESSION["wcs_user_admin"] == 1) { //Wenn Benutzer Admin-Rechte hat
 			"acat_public, acat_struct, acat_template, acat_sort, acat_uid, acat_alias, acat_hidden, acat_topcount, ".
 			"acat_redirect, acat_order, acat_cache, acat_nosearch, acat_nositemap, acat_permit, acat_maxlist, ".
 			"acat_cntpart, acat_pagetitle, acat_paginate, acat_overwrite, acat_archive, acat_class, acat_keywords, ".
-			"acat_cpdefault, acat_lang, acat_lang_type, acat_lang_id) VALUES ('".
+			"acat_cpdefault, acat_lang, acat_lang_type, acat_lang_id, acat_disable301) VALUES ('".
 			getpostvar($_POST["acat_name"])."','".
 			getpostvar($_POST["acat_info"], 32000)."',".
 			(isset($_POST["acat_aktiv"]) ? 1 : 0).",".
@@ -143,7 +144,7 @@ if($_SESSION["wcs_user_admin"] == 1) { //Wenn Benutzer Admin-Rechte hat
 			"'".$acat_permit."', ".intval($_POST["acat_maxlist"]).", '".aporeplace($acat_cntpart)."','".
 			getpostvar($_POST["acat_pagetitle"])."', ".(isset($_POST["acat_paginate"]) ? 1 : 0).", '".getpostvar($_POST["acat_overwrite"])."',".
 			(empty($_POST["acat_archive"]) ? 0 : 1).", '".aporeplace($acat_class)."', '".aporeplace($acat_keywords)."', ".intval($_POST["acat_cpdefault"]).",".
-			_dbEscape($acat_lang).','._dbEscape($acat_lang_type).','._dbEscape($acat_lang_id).')';
+			_dbEscape($acat_lang).','._dbEscape($acat_lang_type).','._dbEscape($acat_lang_id).','.(empty($_POST["acat_disable301"]) ? '0' : '1').')';
 			if($result = mysql_query($sql, $db) or die("MySQL Error: ".mysql_error())) {
 				$ref .= "&cat=".mysql_insert_id($db);
 			}
@@ -187,7 +188,8 @@ if($_SESSION["wcs_user_admin"] == 1) { //Wenn Benutzer Admin-Rechte hat
 				"acat_cpdefault=".intval($_POST["acat_cpdefault"]).','.
 				"acat_lang="._dbEscape($acat_lang).','.
 				"acat_lang_type="._dbEscape($acat_lang_type).','.
-				"acat_lang_id="._dbEscape($acat_lang_id).
+				"acat_lang_id="._dbEscape($acat_lang_id).','.
+				"acat_disable301=".(empty($_POST["acat_disable301"]) ? '0' : '1').
 			" WHERE acat_id=".intval($_POST["acat_id"]);
 		
 			mysql_query($sql, $db) or die(_report_error('DB', $sql));